+++ packages/fe-fpm-writer/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/fe-fpm-writer
 
+## 1.3.14
+
+### Patch Changes
+
+#### Dependency Updates
+
+- update dependencies [open-ux-odata] [[c0a9e6b](https://github.com/SAP/open-ux-tools/commit/c0a9e6b6f81bf1e24a2712e25dde08d72bd12604)]
+
+#### Workspace Updates
+
+- @sap-ux/fiori-annotation-api 1.0.23 → 1.0.24
+- @sap-ux/project-access 2.1.10 → 2.1.11
+
 ## 1.3.13
 
 ### Patch Changes