+++ packages/fiori-annotation-api/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/fiori-annotation-api
 
+## 1.0.24
+
+### Patch Changes
+
+#### Dependency Updates
+
+- update dependencies [open-ux-odata] [[c0a9e6b](https://github.com/SAP/open-ux-tools/commit/c0a9e6b6f81bf1e24a2712e25dde08d72bd12604)]
+
+#### Workspace Updates
+
+- @sap-ux/project-access 2.1.10 → 2.1.11
+- @sap-ux/cds-odata-annotation-converter 1.0.9 → 1.0.9
+
 ## 1.0.23
 
 ### Patch Changes