+++ packages/ui5-application-inquirer/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @sap-ux/ui5-application-inquirer
 
+## 1.0.32
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/inquirer-common 1.0.31 → 1.0.32
+- @sap-ux/project-access 2.1.10 → 2.1.11
+- @sap-ux/project-input-validator 1.0.15 → 1.0.16
+
 ## 1.0.31
 
 ### Patch Changes